+++ plugins/set_value/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+* Support for changing logger level and pattern during execution.
+  [#363](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/363)
 
 ### Changed